requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
beanie==1.10.9
black==23.9.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.89.1
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
lazy-model==0.2.0
loguru==0.7.2
motor==2.5.1
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.4
pydantic_core==2.6.3
PyJWT==2.8.0
pymongo==3.13.0
pyparsing==3.1.1
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2